4.22*: bloglist expansion, step10

+++ part4/blogs/controllers/blogs.js
@@ -1,40 +1,50 @@
 const blogsRouter = require("express").Router()
 const Blog = require("../models/blog")
+const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
 
 // Get all blogs
-blogsRouter.get("/", (request, response) => {
-  Blog
-    .find({})
-    .then(blogs => {
-      response.json(blogs)
-    })
-    .catch(() => {
-      console.log("not succesful")
-    })
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
+
+  response.json(blogs)
 })
 
-// Get specific blog by ID
-blogsRouter.get("/:id", (request, response) => {
-  Blog
-    .findById(request.params.id)
-    .then(blog => {
-      response.json(blog)
-    })
-    .catch(() => {
-      console.log("not succesful")
-    })
+// Get blog by ID
+blogsRouter.get("/:id", async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  response.json(blog)
 })
 
 // Create new Blog
-blogsRouter.post("/", async (request, response) => {
-  const blog = new Blog(request.body)
+blogsRouter.post("/", userExtractor, async (request, response) => {
+  const body = request.body
+
   // Check if missing title or url
-  if (!blog.title || !blog.url) {
+  if (!body.title || !body.url) {
     return response.status(400).json({ error: "Title and URL are required" })
   }
 
-  const result = await blog.save()
-  response.status(201).json(result)
+  // Find user obj associated with token
+  const user = await User.findById(request.user)
+
+  // Create new Blog post
+  const blog = new Blog({
+    title: body.title,
+    url: body.url,
+    likes: body.likes,
+    user: user.id
+  })
+
+  // Save new blog
+  const savedBlog = await blog.save()
+
+  // Append blog to user
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.json(savedBlog)
 })
 
 // Update likes
@@ -48,9 +58,20 @@ blogsRouter.put("/:id", async (request, response) => {
 })
 
 // Delete specific blog by ID
-blogsRouter.delete("/:id", async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
+
+  // Get blog
+  const blog = await Blog.findById(request.params.id)
+
+  // Compare token with blog, ensure only user who created can delete
+  if (!request.user === blog.user.toString())
+  {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  // Delete blog post and return 204
+  await Blog.findByIdAndDelete(request.params.id)
+  response.status(204).end()
 })
 
 module.exports = blogsRouter
